# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Body Boost",
        "description": "Body Boost provides users with access to 1,000 training exercise instructions, offering detailed information to help them perform various exercises effectively.",
        "version": "1.0.0"
    },
    "paths": {
        "/api/register": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Register a new user",
                "description": "Registers a new user and returns an authentication token.",
                "operationId": "registerUser",
                "requestBody": {
                    "description": "User registration data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password",
                                    "phone"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "Full name of the user",
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "email": {
                                        "description": "Email address",
                                        "type": "string",
                                        "format": "email",
                                        "example": "john.doe@example.com"
                                    },
                                    "password": {
                                        "description": "User's password",
                                        "type": "string",
                                        "format": "password",
                                        "example": "SecurePass123!"
                                    },
                                    "phone": {
                                        "description": "User's phone number",
                                        "type": "string",
                                        "example": "+1234567890"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "User registered successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "accepted"
                                        },
                                        "token": {
                                            "type": "string",
                                            "example": "1|aBcDeFgHiJkLmNoPqRsTuVwXyZ"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "Conflict - User phone already exists",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "user phone already exists"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "An error occurred while registering the user."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/login": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Login a user",
                "description": "Authenticates a user and returns an authentication token.",
                "operationId": "loginUser",
                "requestBody": {
                    "description": "User login credentials",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "description": "Email address",
                                        "type": "string",
                                        "format": "email",
                                        "example": "john.doe@example.com"
                                    },
                                    "password": {
                                        "description": "User's password",
                                        "type": "string",
                                        "format": "password",
                                        "example": "SecurePass123!"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User logged in successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "massage": {
                                            "type": "string",
                                            "example": "user logged in"
                                        },
                                        "token": {
                                            "type": "string",
                                            "example": "2|zYxWvUtSrQpOnMlKjIhGfEdCbA"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized - Invalid credentials",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "email or password do not match"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "An error occurred during login."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/bodypart": {
            "post": {
                "tags": [
                    "Exercise"
                ],
                "summary": "Retrieve exercises by body part",
                "description": "Fetches a list of exercises filtered by the specified body part with pagination support.",
                "operationId": "getBodyPartExercises",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "body_part": {
                                        "type": "string",
                                        "example": "chest"
                                    },
                                    "lastId": {
                                        "type": "integer",
                                        "example": -1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Exercises retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "exercise data found"
                                        },
                                        "state": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "Bench Press"
                                                    },
                                                    "body_part": {
                                                        "type": "string",
                                                        "example": "chest"
                                                    },
                                                    "equipment": {
                                                        "type": "string",
                                                        "example": "barbell"
                                                    },
                                                    "target": {
                                                        "type": "string",
                                                        "example": "pectorals"
                                                    },
                                                    "secondary_muscles": {
                                                        "type": "string",
                                                        "example": "triceps, deltoids"
                                                    },
                                                    "instructions": {
                                                        "type": "string",
                                                        "example": "Lie on a bench with a barbell..."
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid body_part or id",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "invalid body_part: chest or id: 5"
                                        },
                                        "state": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "data": {
                                            "type": "string",
                                            "example": ""
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/equipment": {
            "post": {
                "tags": [
                    "Exercise"
                ],
                "summary": "Retrieve exercises by equipment",
                "description": "Fetches a list of exercises filtered by the specified equipment with pagination support.",
                "operationId": "getEquipmentExercises",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "equipment": {
                                        "type": "string",
                                        "example": "dumbbell"
                                    },
                                    "lastId": {
                                        "type": "integer",
                                        "example": -1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Exercises retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "exercise data found"
                                        },
                                        "state": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 2
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "Dumbbell Curl"
                                                    },
                                                    "body_part": {
                                                        "type": "string",
                                                        "example": "biceps"
                                                    },
                                                    "equipment": {
                                                        "type": "string",
                                                        "example": "dumbbell"
                                                    },
                                                    "target": {
                                                        "type": "string",
                                                        "example": "biceps"
                                                    },
                                                    "secondary_muscles": {
                                                        "type": "string",
                                                        "example": "forearms"
                                                    },
                                                    "instructions": {
                                                        "type": "string",
                                                        "example": "Stand with a dumbbell in each hand..."
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid equipment or id",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "invalid equipment: dumbbell or id: 5"
                                        },
                                        "state": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "data": {
                                            "type": "string",
                                            "example": ""
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/target": {
            "post": {
                "tags": [
                    "Exercise"
                ],
                "summary": "Retrieve exercises by target muscle",
                "description": "Fetches a list of exercises filtered by the specified target muscle with pagination support.",
                "operationId": "getTargetExercises",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "target": {
                                        "type": "string",
                                        "example": "biceps"
                                    },
                                    "lastId": {
                                        "type": "integer",
                                        "example": -1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Exercises retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "exercise data found"
                                        },
                                        "state": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 3
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "Hammer Curl"
                                                    },
                                                    "body_part": {
                                                        "type": "string",
                                                        "example": "biceps"
                                                    },
                                                    "equipment": {
                                                        "type": "string",
                                                        "example": "dumbbell"
                                                    },
                                                    "target": {
                                                        "type": "string",
                                                        "example": "biceps"
                                                    },
                                                    "secondary_muscles": {
                                                        "type": "string",
                                                        "example": "forearms"
                                                    },
                                                    "instructions": {
                                                        "type": "string",
                                                        "example": "Hold a dumbbell in each hand..."
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid target or id",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "invalid target: biceps or id: 5"
                                        },
                                        "state": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "data": {
                                            "type": "string",
                                            "example": ""
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter your bearer token in the format **Bearer <token>**",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "tags": [
        {
            "name": "User",
            "description": "Operations related to user authentication and account management"
        },
        {
            "name": "Exercise",
            "description": "Operations related to exercise data retrieval"
        }
    ]
}